+++ web/src/assets/styles/variables.scss
@@ -1,0 +1,4 @@
+$bg-body: #0b3d91;
+$white: #fff;
+$nasa-orange: #fc3d21;
+$font-family-inter: 'Inter', sans-serif;
