+++ web/src/component/ShowMedia.jsx
@@ -1,29 +1,37 @@
-// /* eslint-disable array-callback-return */
-// import { useEffect } from 'react';
-// import { useDispatch, useSelector } from 'react-redux';
+/* eslint-disable react/prop-types */
+/* eslint-disable no-unused-vars */
+import React, { useState, useEffect } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
+import { v4 as uuidv4 } from 'uuid';
+import { fetchPhotos } from '../redux/media/media';
 
-// // import NavBar from '../navbar/NavBar';
-// import Photo from './Photo';
-// import { fetchPhotos } from '../redux/media/media';
+const ShowMedia = () => {
+  const photos = useSelector((state) => state.photosReducer);
+  const dispatch = useDispatch();
 
-// const ShowMedia = () => {
-//   const movies = useSelector((state) => state.moviesReducer.movies);
-//   const dispatch = useDispatch();
+  useEffect(() => {
+    dispatch(fetchPhotos());
+  }, []);
 
-//   useEffect(() => {
-//     dispatch(fetchPhotos());
-//   }, []);
+  return (
+    <>
+      <h1 className="text-center font-extrabold text-[35px]">Welcome to Nasa Media Library</h1>
+      {photos.photos.length > 0 && photos.photos.map((photo) => (
+        <div
+          key={uuidv4()}
+        >
+          <div className="p-10 flex flex-col">
+            <div className="grid grid-cols-4">
+              <img src={photo.links[0].href} alt={photo.data[0].title} className="w-[300px] h-[300px]" />
+            </div>
+            <p>{photo.data[0].location}</p>
+            <p>{photo.data[0].date_created}</p>
+            <p>{photo.data[0].description}</p>
+          </div>
+        </div>
+      ))}
+    </>
+  );
+};
 
-//   return (
-//     <>
-//       {/* <NavBar /> */}
-//       <div className="container mx-auto grid grid-cols-1 md:grid-cols-3
-//  gap-8 py-24 pl-20 md:pl-32 pr-4">
-//         {movies
-//           && movies.map((photo) => (<Photo key={photo.id} photo={photo} />))}
-//       </div>
-//     </>
-//   );
-// };
-
-// export default ShowMedia;
+export default ShowMedia;
