+++ web/src/index.scss
@@ -3,9 +3,17 @@
 @tailwind components;
 @tailwind utilities;
 
+@import url('https://fonts.googleapis.com/css2?family=Inter:wght@100;200;300;400;500;600;700;800;900&family=Lato&family=Mochiy+Pop+P+One&family=Montserrat:wght@700&family=Nunito:ital,wght@0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&family=Open+Sans:wght@400;600&family=Roboto+Slab:wght@300;700&family=Roboto:wght@300&family=Rubik:ital,wght@0,300;0,400;0,500;0,600;0,700;0,800;1,300;1,400;1,500;1,600;1,700;1,800&family=Space+Grotesk:wght@300;400;500;600;700&display=swap');
+@import './assets/styles/variables.scss';
+
 *,
 ::after,
 ::before {
   margin: 0;
   padding: 0;
+}
+
+body {
+  background: $bg-body;
+  color: $white;
 }
