+++ web/src/pages/SearchPage.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Home from '../component/Home';
+// import Home from '../component/Home';
 import Search from '../component/Search';
 // import ShowMedia from '../component/ShowMedia';
 
@@ -7,7 +7,7 @@ const SearchPage = () => (
   <div>
     <Search />
     {/* <ShowMedia /> */}
-    <Home />
+    {/* <Home /> */}
   </div>
 );
